:sparkles: Display co-travellers on status detail (#599)
closes #580

Reviewed-on: https://codeberg.org/traewelldroid/traewelldroid/pulls/599
Co-committed-by: Jonas H. <jonas@hbch.de>

+++ app/src/main/kotlin/de/hbch/traewelling/api/models/status/Journey.kt
@@ -8,7 +8,7 @@ import de.hbch.traewelling.api.models.trip.ProductType
 import java.time.ZonedDateTime
 
 data class Journey(
-    @SerializedName("trip") val tripId: Int,
+    @SerializedName("trip") val tripId: ULong,
     @SerializedName("hafasId") val hafasTripId: String,
     val category: ProductType?,
     @SerializedName("lineName") val line: String,
